Remove django-extensions
It's an extra dependency which needs installing in the fundamental
django installation, even though it's mainly convenience
functions for local development.

I've never used any of its convenience functions, and I don't think
anyone else has for a long time. So I'm making a call that should
just stick to core Django functions in this project for the time
being.